Split MCP server module
